+++ controllers/userController.js
@@ -1,4 +1,6 @@
 const User = require("../models/User");
+const Blog = require("../models/Blog"); 
+const { deleteFile } = require("../middleware/upload");
 
 const getProfile = (req, res) => {
   res.render("profile", {
@@ -19,6 +21,9 @@ const updateProfile = async (req, res) => {
         user.password = password;
       }
       if (req.file) {
+        if (user.profilePic) {
+          deleteFile(user.profilePic);
+        }
         user.profilePic = req.file.filename;
       }
       await user.save();
@@ -37,9 +42,27 @@ const updateProfile = async (req, res) => {
 };
 
 
+
 const deleteAccount = async (req, res) => {
   try {
-    await User.deleteOne({ email: req.session.user.email });
+    const user = await User.findOne({ email: req.session.user.email });
+
+    if (user) {
+      if (user.profilePic) {
+        deleteFile(user.profilePic);
+      }
+      const userBlogs = await Blog.find({ author: user._id });
+      userBlogs.forEach(blog => {
+        if (blog.coverImage) {
+          deleteFile(blog.coverImage);
+        }
+      });
+
+
+      await Blog.deleteMany({ author: user._id });
+      await User.deleteOne({ email: req.session.user.email });
+    }
+
     req.session.destroy((err) => {
       if (err) console.error(err);
       res.redirect("/signup");
@@ -49,7 +72,6 @@ const deleteAccount = async (req, res) => {
     res.status(500).send("Something went wrong. Please try again.");
   }
 };
-
 module.exports = {
   getProfile,
   updateProfile,
